Move utils functions only used in tests to head.js
