+++ src/test/mochitest/head.js
@@ -22,6 +22,7 @@ var { gDevTools } = require("devtools/client/framework/devtools");
 var { BrowserLoader } = Cu.import("resource://devtools/client/shared/browser-loader.js", {});
 var flags = require("devtools/shared/flags");
 var { Task } = require("devtools/shared/task");
+const nodeConstants = require("devtools/shared/dom-node-constants");
 var DevToolsUtils = require("devtools/shared/DevToolsUtils");
 
 flags.testing = true;
@@ -33,6 +34,7 @@ var { require: browserRequire } = BrowserLoader({
 let ReactDOM = browserRequire("devtools/client/shared/vendor/react-dom");
 let React = browserRequire("devtools/client/shared/vendor/react");
 var TestUtils = React.addons.TestUtils;
+const { getGripPreviewItems } = browserRequire("devtools/client/shared/components/reps/reps");
 
 function renderComponent(component, props) {
   const el = React.createElement(component, props, {});
@@ -68,4 +70,50 @@ function testRepRenderModes(modeTests, testName, componentUnderTest, gripStub,
     );
     is(rendered.textContent, expectedOutput, message);
   });
+}
+
+/**
+ * Get an array of all the items from the grip in parameter (including the grip itself)
+ * which can be selected in the inspector.
+ *
+ * @param {Object} Grip
+ * @return {Array} Flat array of the grips which can be selected in the inspector
+ */
+function getSelectableInInspectorGrips(grip) {
+  let grips = new Set(getFlattenedGrips([grip]));
+  return [...grips].filter(isGripSelectableInInspector);
+}
+
+/**
+ * Indicate if a Grip can be selected in the inspector,
+ * i.e. if it represents a node element.
+ *
+ * @param {Object} Grip
+ * @return {Boolean}
+ */
+function isGripSelectableInInspector(grip) {
+  return grip
+    && typeof grip === "object"
+    && grip.preview
+    && [
+      nodeConstants.TEXT_NODE,
+      nodeConstants.ELEMENT_NODE
+    ].includes(grip.preview.nodeType);
+}
+
+/**
+ * Get a flat array of all the grips and their preview items.
+ *
+ * @param {Array} Grips
+ * @return {Array} Flat array of the grips and their preview items
+ */
+function getFlattenedGrips(grips) {
+  return grips.reduce((res, grip) => {
+    let previewItems = getGripPreviewItems(grip);
+    let flatPreviewItems = previewItems.length > 0
+      ? getFlattenedGrips(previewItems)
+      : [];
+
+    return [...res, grip, ...flatPreviewItems];
+  }, []);
 }
