+++ src/test/mochitest/test_reps_event.html
@@ -21,7 +21,6 @@ window.onload = Task.async(function* () {
     REPS,
     MODE,
     getRep,
-    getSelectableInInspectorGrips,
   } = browserRequire("devtools/client/shared/components/reps/reps");
   let { Event } = REPS;
