+++ src/test/mochitest/test_reps_grip-array.html
@@ -21,7 +21,6 @@ window.onload = Task.async(function* () {
     REPS,
     MODE,
     getRep,
-    getSelectableInInspectorGrips,
   } = browserRequire("devtools/client/shared/components/reps/reps");
   let { GripArray } = REPS;
