+++ src/test/mochitest/test_reps_grip-map.html
@@ -23,7 +23,6 @@ window.onload = Task.async(function* () {
     REPS,
     MODE,
     getRep,
-    getSelectableInInspectorGrips,
   } = browserRequire("devtools/client/shared/components/reps/reps");
   let { GripMap } = REPS;
