+++ src/test/mochitest/test_reps_promise.html
@@ -23,7 +23,6 @@ window.onload = Task.async(function* () {
     REPS,
     MODE,
     getRep,
-    getSelectableInInspectorGrips,
   } = browserRequire("devtools/client/shared/components/reps/reps");
   let { PromiseRep } = REPS;
